ml_service/requirements.txt
django==5.0.4
djangorestframework==3.15.1
python-dotenv==1.0.1
requests==2.31.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.3